fix: prevent file traversal using helm file values param and application details api (#8606)
* fix: prevent file traversal using helm file values param and application details api


* apply reviewer notes: move resolve.go into separate package; use uuid to generate random file
