+++ reposerver/repository/repository.proto
@@ -83,6 +83,7 @@ message RepoServerAppDetailsQuery {
     string appName = 5;
     bool noCache = 6;
     bool noRevisionCache = 7;
+    github.com.argoproj.argo_cd.v2.pkg.apis.application.v1alpha1.HelmOptions helmOptions = 10;
 }
 
 // RepoAppDetailsResponse application details
